Rename public/ folder
Follow the standard Ruby structure of lib/gem_name

This also helps us to understand the difference between
resource_registry and schema_registry at first sight
