+++ lib/resource_registry.rb
@@ -3,15 +3,15 @@
 
 require "sorbet-runtime"
 
-require "public/configuration"
-require "public/resource"
-require "public/versions"
-require "public/versions/version"
-require "public/entity_finder"
-require "public/resource_struct_builder"
-require "public/registry"
-require "public/serializer"
-require "public/repositories/base"
+require "resource_registry/configuration"
+require "resource_registry/resource"
+require "resource_registry/versions"
+require "resource_registry/versions/version"
+require "resource_registry/entity_finder"
+require "resource_registry/resource_struct_builder"
+require "resource_registry/registry"
+require "resource_registry/serializer"
+require "resource_registry/repositories/base"
 require "schema_registry/registry"
 require "schema_registry/json_schema_mapper"
 require "schema_registry/maybe"
