+++ lib/resource_registry/capabilities/capability_config.rb
