+++ lib/resource_registry/capability_factory.rb
