+++ lib/resource_registry/configuration.rb
