+++ lib/resource_registry/entity_finder.rb
