+++ lib/resource_registry/load_resources_from_cache.rb
