+++ lib/resource_registry/registry.rb
