+++ lib/resource_registry/relationship.rb
