+++ lib/resource_registry/relationship_type.rb
