+++ lib/resource_registry/relationship_type_factory.rb
