+++ lib/resource_registry/relationship_types/belongs_to.rb
