+++ lib/resource_registry/relationship_types/has_many_through.rb
