+++ lib/resource_registry/relationship_types/has_one.rb
