+++ lib/resource_registry/repositories/base.rb
