+++ lib/resource_registry/resource.rb
