+++ lib/resource_registry/resource_struct_builder.rb
