+++ lib/resource_registry/serializer.rb
