+++ lib/resource_registry/verb.rb
