+++ lib/resource_registry/versions.rb
