+++ lib/resource_registry/versions/version.rb
